JWT/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
sqlparse==0.4.4
typing-extensions==4.6.3
requests==2.31.0
